# Respect bufferSize on SocketHttpHandler's ContentLengthReadStream