@stephentoub 2018-10-15
@dotnet-bot test CROSS Check please ("The remote name could not be resolved")