@stephentoub 2018-10-15
@dotnet-bot test Windows x64 Debug Build please
@dotnet-bot test Linux x64 Release Build please